+++ src/elements/Icon/Icon.jsx
@@ -18,30 +18,38 @@ const Icon = React.createClass({
         ]),
     },
 
-    // cononical icons
-    // TODO: some icons are broken, probably because of bug in material icons release
-    // @see https://github.com/google/material-design-icons/issues/311
+    // cononical icon names
     canonicalIcons: {
         'edit': 'mode_edit',
-        'delete': 'delete',
+        'remove': 'delete',
+        'save': 'save',
+        'confirm': 'check',
+        'cancel': 'close',
         'arrow_nextpage': 'navigate_next',
         'arrow_prevpage': 'navigate_before',
-        'arrow_lastpage': 'skip_next', // TODO: workaround for 'last_page',
-        'arrow_firstpage': 'skip_previous', // TODO: workaround for 'first_page',
+        'arrow_lastpage': 'last_page',
+        'arrow_firstpage': 'first_page',
         'arrow_dropdown': 'arrow_drop_down',
         'expand_more': 'expand_more',
         'expand_less': 'expand_less',
         'menu_more': 'more_vert',
+        'adjust': 'tune',
         'filter': 'filter_list',
         'sort': 'swap_vert',
         'hide': 'close',
+        'settings': 'settings',
         'access_forbidden': 'lock_outline',
+        'delete': 'delete', // 'delete' is deprecated
     },
 
+    // fallback tooltips for all canonocal icon names
     // TODO: use translations
     canonicalTooltips: {
         'edit': 'Edit',
-        'delete': 'Delete',
+        'remove': 'Remove',
+        'save': 'Save',
+        'confirm': 'Confirm',
+        'cancel': 'Cancel',
         'arrow_nextpage': 'Next page',
         'arrow_prevpage': 'Previous page',
         'arrow_lastpage': 'Last page',
@@ -50,10 +58,13 @@ const Icon = React.createClass({
         'expand_more': 'Show more',
         'expand_less': 'Show less',
         'menu_more': 'Open menu',
+        'adjust': 'Adjust settings',
         'filter': 'Filter data',
         'sort': 'Sort data',
         'hide': 'Hide',
+        'settings': 'Administrate settings',
         'access_forbidden': 'No write access',
+        'delete': 'Remove', // 'delete' is deprecated
     },
 
     // template rendering
@@ -74,13 +85,25 @@ const Icon = React.createClass({
 
         if (typeof this.canonicalIcons[name] !== 'undefined') {
             name = this.canonicalIcons[name];
+            if (name === 'delete' && __DEBUG__) {
+                console.warn('Do not us "delete" as icon name. Please use "remove" instead.');
+            }
         } else {
-            // TODO: add warning about usage of non-canonical icons
+            if (__DEBUG__) {
+                console.log(
+                    '%cFound usage of "' + name + '" as icon name. Use canonical icon names if possible.',
+                    'color: orange'
+                );
+            }
         }
         if (typeof ligaturcodes[name] !== 'undefined') {
             name = ligaturcodes[name];
         } else {
-            // TODO: add error about unknown ligature code/icon name
+            if (__DEBUG__) {
+                console.error(
+                    '"' + name + '" is not a valid icon name.'
+                );
+            }
         }
 
         const classes = classNames('material-icons', className);