+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [2.21.0] 2017-11-02
+
+### Fixed
+- fix the fallback tooltip for the delete/remove icon to "Remove."
+
+### Added
+- insert canonical icons and fallback tooltips for 'save', 'confirm', 'cancel', 'adjust' and 'settings'
+
 ## [2.20.1] 2017-10-16
 
 ### Fixed